update train bash script fine

+++ transcribe/test.py
@@ -4,12 +4,12 @@ from flax.jax_utils import replicate
 from flax.training.common_utils import shard
 from jax import device_get, pmap
 from transformers import WhisperProcessor
-from whisper_jax import FlaxWhisperForConditionalGeneration
+from transformers import FlaxWhisperForConditionalGeneration
 
 # load the processor and model
 processor = WhisperProcessor.from_pretrained("openai/whisper-large-v2")
 model, params = FlaxWhisperForConditionalGeneration.from_pretrained(
-    "openai/whisper-large-v2", dtype=jnp.float32, _do_init=False,
+    "openai/whisper-large-v2", dtype=jnp.bfloat16, _do_init=False,
 )
 
 def generate_fn(input_features):
